Refactor: Rename `SubsidiaryFeeExtensions` to `SubsidiaryFeeHelper`
