+++ src/EPR.RegulatorService.Frontend.Web/Extensions/ProducerPaymentResponseExtensions.cs
@@ -1,11 +1,12 @@
 namespace EPR.RegulatorService.Frontend.Web.Extensions;
 
 using EPR.RegulatorService.Frontend.Core.Models.RegistrationSubmissions;
+using EPR.RegulatorService.Frontend.Web.Helpers;
 
 internal static class ProducerPaymentResponseExtensions
 {
     internal static decimal GetNetSubsidiaryCompaniesFee(this ProducerPaymentResponse response) =>
-        SubsidiaryFeeExtensions.GetNetSubsidiaryCompaniesFee(
+        SubsidiaryFeeHelper.GetNetSubsidiaryCompaniesFee(
             response.SubsidiaryFee,
             response.SubsidiariesFeeBreakdown);
 }
